backend/requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.38.0
psycopg2-binary==2.9.11
numpy==2.3.4
scipy==1.16.2
python-dotenv==1.1.1
pydantic==2.12.3
